백준 2343 재시도 230727 (#1)
- 블로그 보고 재시도 -> 통과
